Revert "[GSM Observability] Metadata Exchange Implementation (#34051)"
This reverts commit 7c79712d13285b640a7f960c7d2d2586cb9f52e9.
