+++ src/core/lib/transport/metadata_batch.h
@@ -305,13 +305,6 @@ struct GrpcTagsBinMetadata : public SimpleSliceBasedMetadata {
   static absl::string_view key() { return "grpc-tags-bin"; }
 };
 
-// XEnvoyPeerMetadata
-struct XEnvoyPeerMetadata : public SimpleSliceBasedMetadata {
-  static constexpr bool kRepeatable = false;
-  using CompressionTraits = StableValueCompressor;
-  static absl::string_view key() { return "x-envoy-peer-metadata"; }
-};
-
 // :authority metadata trait.
 struct HttpAuthorityMetadata : public SimpleSliceBasedMetadata {
   static constexpr bool kRepeatable = false;
@@ -1481,7 +1474,6 @@ using grpc_metadata_batch_base = grpc_core::MetadataMap<
     grpc_core::GrpcServerStatsBinMetadata, grpc_core::GrpcTraceBinMetadata,
     grpc_core::GrpcTagsBinMetadata, grpc_core::GrpcLbClientStatsMetadata,
     grpc_core::LbCostBinMetadata, grpc_core::LbTokenMetadata,
-    grpc_core::XEnvoyPeerMetadata,
     // Non-encodable things
     grpc_core::GrpcStreamNetworkState, grpc_core::PeerString,
     grpc_core::GrpcStatusContext, grpc_core::GrpcStatusFromWire,
