+++ src/cpp/ext/gsm/BUILD
@@ -32,32 +32,21 @@ grpc_cc_library(
     name = "gsm_observability",
     srcs = [
         "gsm_observability.cc",
-        "metadata_exchange.cc",
     ],
     hdrs = [
         "gsm_observability.h",
-        "metadata_exchange.h",
     ],
     external_deps = [
         "absl/container:flat_hash_set",
-        "absl/meta:type_traits",
         "absl/status",
         "absl/status:statusor",
         "absl/strings",
-        "absl/types:optional",
-        "absl/types:variant",
         "otel/sdk/src/metrics",
-        "otel/sdk:headers",
-        "upb_lib",
     ],
     language = "c++",
     visibility = ["//:__subpackages__"],
     deps = [
         "//:gpr_platform",
-        "//:grpc_base",
-        "//:protobuf_struct_upb",
-        "//src/core:env",
-        "//src/core:slice",
         "//src/cpp/ext/otel:otel_plugin",
     ],
 )
