+++ src/cpp/ext/otel/BUILD
@@ -36,7 +36,6 @@ grpc_cc_library(
         "otel_server_call_tracer.cc",
     ],
     hdrs = [
-        "key_value_iterable.h",
         "otel_call_tracer.h",
         "otel_client_filter.h",
         "otel_plugin.h",
@@ -45,13 +44,13 @@ grpc_cc_library(
     external_deps = [
         "absl/base:core_headers",
         "absl/container:flat_hash_set",
+        "absl/container:inlined_vector",
         "absl/status",
         "absl/status:statusor",
         "absl/strings",
         "absl/strings:str_format",
         "absl/time",
         "absl/types:optional",
-        "absl/types:span",
         "otel/api",
     ],
     language = "c++",
@@ -63,7 +62,6 @@ grpc_cc_library(
         "//:gpr_platform",
         "//:grpc_base",
         "//:grpc_client_channel",
-        "//:grpc_public_hdrs",
         "//:legacy_context",
         "//src/core:arena",
         "//src/core:arena_promise",
