+++ src/cpp/ext/otel/otel_client_filter.cc
@@ -20,13 +20,13 @@
 
 #include "src/cpp/ext/otel/otel_client_filter.h"
 
-#include <array>
 #include <functional>
 #include <initializer_list>
 #include <memory>
 #include <string>
 #include <utility>
 
+#include "absl/container/inlined_vector.h"
 #include "absl/status/status.h"
 #include "absl/strings/str_format.h"
 #include "absl/strings/string_view.h"
@@ -34,25 +34,21 @@
 #include "absl/time/clock.h"
 #include "absl/time/time.h"
 #include "absl/types/optional.h"
-#include "absl/types/span.h"
 #include "opentelemetry/context/context.h"
 #include "opentelemetry/metrics/sync_instruments.h"
 
-#include <grpc/status.h>
 #include <grpc/support/log.h>
 #include <grpc/support/time.h>
 
 #include "src/core/ext/filters/client_channel/client_channel.h"
 #include "src/core/lib/channel/channel_stack.h"
 #include "src/core/lib/channel/context.h"
-#include "src/core/lib/channel/status_util.h"
 #include "src/core/lib/gprpp/sync.h"
 #include "src/core/lib/promise/context.h"
 #include "src/core/lib/resource_quota/arena.h"
 #include "src/core/lib/slice/slice.h"
 #include "src/core/lib/slice/slice_buffer.h"
 #include "src/core/lib/transport/metadata_batch.h"
-#include "src/cpp/ext/otel/key_value_iterable.h"
 #include "src/cpp/ext/otel/otel_call_tracer.h"
 #include "src/cpp/ext/otel/otel_plugin.h"
 
@@ -104,34 +100,11 @@ OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::
     : parent_(parent),
       arena_allocated_(arena_allocated),
       start_time_(absl::Now()) {
-  // We don't have the peer labels at this point.
-  if (OTelPluginState().labels_injector != nullptr) {
-    local_labels_ = OTelPluginState().labels_injector->GetLocalLabels();
-  }
+  // TODO(yashykt): Figure out how to get this to work with absl::string_view
   if (OTelPluginState().client.attempt.started != nullptr) {
-    std::array<std::pair<absl::string_view, absl::string_view>, 2>
-        additional_labels = {
-            {{OTelMethodKey(),
-              absl::StripPrefix(parent_->path_.as_string_view(), "/")},
-             {OTelTargetKey(), parent_->parent_->target()}}};
-    KeyValueIterable labels(local_labels_.get(), peer_labels_.get(),
-                            additional_labels);
-    OTelPluginState().client.attempt.started->Add(1, labels);
-  }
-}
-
-void OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::
-    RecordReceivedInitialMetadata(grpc_metadata_batch* recv_initial_metadata) {
-  if (OTelPluginState().labels_injector != nullptr) {
-    peer_labels_ =
-        OTelPluginState().labels_injector->GetPeerLabels(recv_initial_metadata);
-  }
-}
-
-void OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::
-    RecordSendInitialMetadata(grpc_metadata_batch* send_initial_metadata) {
-  if (OTelPluginState().labels_injector != nullptr) {
-    OTelPluginState().labels_injector->AddLabels(send_initial_metadata);
+    OTelPluginState().client.attempt.started->Add(
+        1, {{std::string(OTelMethodKey()), std::string(parent_->method_)},
+            {std::string(OTelTargetKey()), parent_->parent_->target()}});
   }
 }
 
@@ -165,19 +138,13 @@ void OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::
     RecordReceivedTrailingMetadata(
         absl::Status status, grpc_metadata_batch* /*recv_trailing_metadata*/,
         const grpc_transport_stream_stats* transport_stream_stats) {
-  std::array<std::pair<absl::string_view, absl::string_view>, 3>
-      additional_labels = {
-          {{OTelMethodKey(),
-            absl::StripPrefix(parent_->path_.as_string_view(), "/")},
-           {OTelTargetKey(), parent_->parent_->target()},
-           {OTelStatusKey(),
-            grpc_status_code_to_string(
-                static_cast<grpc_status_code>(status.code()))}}};
-  KeyValueIterable labels(local_labels_.get(), peer_labels_.get(),
-                          additional_labels);
+  absl::InlinedVector<std::pair<std::string, std::string>, 2> attributes = {
+      {std::string(OTelMethodKey()), std::string(parent_->method_)},
+      {std::string(OTelStatusKey()), absl::StatusCodeToString(status.code())},
+      {std::string(OTelTargetKey()), parent_->parent_->target()}};
   if (OTelPluginState().client.attempt.duration != nullptr) {
     OTelPluginState().client.attempt.duration->Record(
-        absl::ToDoubleSeconds(absl::Now() - start_time_), labels,
+        absl::ToDoubleSeconds(absl::Now() - start_time_), attributes,
         opentelemetry::context::Context{});
   }
   if (OTelPluginState().client.attempt.sent_total_compressed_message_size !=
@@ -186,7 +153,7 @@ void OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::
         transport_stream_stats != nullptr
             ? transport_stream_stats->outgoing.data_bytes
             : 0,
-        labels, opentelemetry::context::Context{});
+        attributes, opentelemetry::context::Context{});
   }
   if (OTelPluginState().client.attempt.rcvd_total_compressed_message_size !=
       nullptr) {
@@ -194,7 +161,7 @@ void OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::
         transport_stream_stats != nullptr
             ? transport_stream_stats->incoming.data_bytes
             : 0,
-        labels, opentelemetry::context::Context{});
+        attributes, opentelemetry::context::Context{});
   }
 }
 
@@ -227,7 +194,10 @@ void OpenTelemetryCallTracer::OpenTelemetryCallAttemptTracer::RecordAnnotation(
 OpenTelemetryCallTracer::OpenTelemetryCallTracer(
     OpenTelemetryClientFilter* parent, grpc_core::Slice path,
     grpc_core::Arena* arena)
-    : parent_(parent), path_(std::move(path)), arena_(arena) {}
+    : parent_(parent),
+      path_(std::move(path)),
+      method_(absl::StripPrefix(path_.as_string_view(), "/")),
+      arena_(arena) {}
 
 OpenTelemetryCallTracer::~OpenTelemetryCallTracer() {}
