+++ src/cpp/ext/otel/otel_client_filter.h
@@ -25,7 +25,6 @@
 #include <utility>
 
 #include "absl/status/statusor.h"
-#include "absl/strings/string_view.h"
 
 #include "src/core/lib/channel/channel_args.h"
 #include "src/core/lib/channel/channel_fwd.h"
@@ -48,7 +47,7 @@ class OpenTelemetryClientFilter : public grpc_core::ChannelFilter {
       grpc_core::CallArgs call_args,
       grpc_core::NextPromiseFactory next_promise_factory) override;
 
-  absl::string_view target() const { return target_; }
+  const std::string& target() const { return target_; }
 
  private:
   explicit OpenTelemetryClientFilter(std::string target)
