+++ src/cpp/ext/otel/otel_plugin.h
@@ -21,63 +21,19 @@
 
 #include <grpc/support/port_platform.h>
 
-#include <stddef.h>
 #include <stdint.h>
 
 #include <memory>
 #include <string>
-#include <utility>
 
 #include "absl/container/flat_hash_set.h"
 #include "absl/strings/string_view.h"
-#include "absl/types/optional.h"
 #include "opentelemetry/metrics/meter_provider.h"
 #include "opentelemetry/metrics/sync_instruments.h"
 #include "opentelemetry/nostd/shared_ptr.h"
 
-#include "src/core/lib/transport/metadata_batch.h"
-
 namespace grpc {
 namespace internal {
-
-// An iterable container interface that can be used as a return type for the
-// OTel plugin's label injector.
-class LabelsIterable {
- public:
-  virtual ~LabelsIterable() = default;
-
-  // Returns the key-value label at the current position or absl::nullopt if the
-  // iterator has reached the end.
-  virtual absl::optional<std::pair<absl::string_view, absl::string_view>>
-  Next() = 0;
-
-  virtual size_t Size() const = 0;
-
-  // Resets position of iterator to the start.
-  virtual void ResetIteratorPosition() = 0;
-};
-
-// An interface that allows you to add additional labels on the calls traced
-// through the OTel plugin.
-class LabelsInjector {
- public:
-  virtual ~LabelsInjector() {}
-  // Read the incoming initial metadata to get the set of labels to be added to
-  // metrics. (Does not include the local labels.)
-  virtual std::unique_ptr<LabelsIterable> GetPeerLabels(
-      grpc_metadata_batch* incoming_initial_metadata) = 0;
-
-  // Get the local labels to be added to metrics. To be used when the peer
-  // metadata is not available, for example, for started RPCs metric.
-  // It is the responsibility of the implementation to make sure that the
-  // backing store for the absl::string_view remains valid for the lifetime of
-  // gRPC.
-  virtual std::unique_ptr<LabelsIterable> GetLocalLabels() = 0;
-
-  // Modify the outgoing initial metadata with metadata information to be sent
-  // to the peer.
-  virtual void AddLabels(grpc_metadata_batch* outgoing_initial_metadata) = 0;
-};
 
 struct OTelPluginState {
   struct Client {
@@ -102,7 +58,6 @@ struct OTelPluginState {
   } server;
   opentelemetry::nostd::shared_ptr<opentelemetry::metrics::MeterProvider>
       meter_provider;
-  std::unique_ptr<LabelsInjector> labels_injector;
 };
 
 const struct OTelPluginState& OTelPluginState();
@@ -135,10 +90,6 @@ class OpenTelemetryPluginBuilder {
   OpenTelemetryPluginBuilder& DisableMetrics(
       const absl::flat_hash_set<absl::string_view>& metric_names);
   // Builds and registers the OTel Plugin
-
-  OpenTelemetryPluginBuilder& SetLabelsInjector(
-      std::unique_ptr<LabelsInjector> labels_injector);
-
   void BuildAndRegisterGlobal();
 
   // The base set of metrics -
@@ -154,7 +105,6 @@ class OpenTelemetryPluginBuilder {
 
  private:
   std::shared_ptr<opentelemetry::metrics::MeterProvider> meter_provider_;
-  std::unique_ptr<LabelsInjector> labels_injector_;
   absl::flat_hash_set<std::string> metrics_;
 };
