+++ test/cpp/ext/gsm/gsm_observability_test.cc
@@ -21,7 +21,6 @@
 #include "google/cloud/opentelemetry/resource_detector.h"
 #include "gtest/gtest.h"
 
-#include "src/core/lib/gprpp/env.h"
 #include "test/core/util/test_config.h"
 
 namespace grpc {
