+++ test/cpp/ext/otel/otel_plugin_test.cc
@@ -31,7 +31,6 @@
 #include "src/core/lib/config/core_configuration.h"
 #include "test/core/util/test_config.h"
 #include "test/cpp/end2end/test_service_impl.h"
-#include "test/cpp/ext/otel/otel_test_library.h"
 
 namespace grpc {
 namespace testing {
@@ -45,8 +44,120 @@ TEST(OTelPluginBuildTest, SdkDependency) {
   opentelemetry::sdk::metrics::MeterProvider();
 }
 
+class MockMetricReader : public opentelemetry::sdk::metrics::MetricReader {
+ public:
+  opentelemetry::sdk::metrics::AggregationTemporality GetAggregationTemporality(
+      opentelemetry::sdk::metrics::InstrumentType) const noexcept override {
+    return opentelemetry::sdk::metrics::AggregationTemporality::kDelta;
+  }
+
+  bool OnForceFlush(std::chrono::microseconds) noexcept override {
+    return true;
+  }
+
+  bool OnShutDown(std::chrono::microseconds) noexcept override { return true; }
+
+  void OnInitialized() noexcept override {}
+};
+
+class OTelPluginEnd2EndTest : public ::testing::Test {
+ protected:
+  using ::testing::Test::SetUp;
+  void SetUp(const absl::flat_hash_set<absl::string_view>& metric_names,
+             bool test_no_meter_provider = false) {
+    // We are resetting the MeterProvider and OpenTelemetry plugin at the start
+    // of each test to avoid test results from one test carrying over to another
+    // test. (Some measurements can get arbitrarily delayed.)
+    grpc_core::CoreConfiguration::Reset();
+    grpc::internal::OpenTelemetryPluginBuilder ot_builder;
+    ot_builder.EnableMetrics(metric_names);
+    if (!test_no_meter_provider) {
+      auto meter_provider =
+          std::make_shared<opentelemetry::sdk::metrics::MeterProvider>();
+      reader_.reset(new grpc::testing::MockMetricReader);
+      meter_provider->AddMetricReader(reader_);
+      ot_builder.SetMeterProvider(std::move(meter_provider));
+    }
+    ot_builder.BuildAndRegisterGlobal();
+    grpc_init();
+    grpc::ServerBuilder builder;
+    int port;
+    // Use IPv4 here because it's less flaky than IPv6 ("[::]:0") on Travis.
+    builder.AddListeningPort("0.0.0.0:0", grpc::InsecureServerCredentials(),
+                             &port);
+    builder.RegisterService(&service_);
+    server_ = builder.BuildAndStart();
+    ASSERT_NE(nullptr, server_);
+    ASSERT_NE(0, port);
+    server_address_ = absl::StrCat("localhost:", port);
+    canonical_server_address_ = absl::StrCat("dns:///", server_address_);
+
+    stub_ = EchoTestService::NewStub(grpc::CreateChannel(
+        server_address_, grpc::InsecureChannelCredentials()));
+  }
+
+  void TearDown() override {
+    server_->Shutdown();
+    grpc_shutdown_blocking();
+    delete grpc_core::ServerCallTracerFactory::Get(grpc_core::ChannelArgs());
+    grpc_core::ServerCallTracerFactory::RegisterGlobal(nullptr);
+  }
+
+  void ResetStub(std::shared_ptr<Channel> channel) {
+    stub_ = EchoTestService::NewStub(channel);
+  }
+
+  void SendRPC() {
+    EchoRequest request;
+    request.set_message("foo");
+    EchoResponse response;
+    grpc::ClientContext context;
+    grpc::Status status = stub_->Echo(&context, request, &response);
+  }
+
+  absl::flat_hash_map<
+      std::string,
+      std::vector<opentelemetry::sdk::metrics::PointDataAttributes>>
+  ReadCurrentMetricsData(
+      absl::AnyInvocable<
+          bool(const absl::flat_hash_map<
+               std::string,
+               std::vector<opentelemetry::sdk::metrics::PointDataAttributes>>&)>
+          continue_predicate) {
+    absl::flat_hash_map<
+        std::string,
+        std::vector<opentelemetry::sdk::metrics::PointDataAttributes>>
+        data;
+    auto deadline = absl::Now() + absl::Seconds(5);
+    do {
+      reader_->Collect([&](opentelemetry::sdk::metrics::ResourceMetrics& rm) {
+        for (const opentelemetry::sdk::metrics::ScopeMetrics& smd :
+             rm.scope_metric_data_) {
+          for (const opentelemetry::sdk::metrics::MetricData& md :
+               smd.metric_data_) {
+            for (const opentelemetry::sdk::metrics::PointDataAttributes& dp :
+                 md.point_data_attr_) {
+              data[md.instrument_descriptor.name_].push_back(dp);
+            }
+          }
+        }
+        return true;
+      });
+    } while (continue_predicate(data) && deadline > absl::Now());
+    return data;
+  }
+
+  const absl::string_view kMethodName = "grpc.testing.EchoTestService/Echo";
+  std::shared_ptr<opentelemetry::sdk::metrics::MetricReader> reader_;
+  std::string server_address_;
+  std::string canonical_server_address_;
+  CallbackTestServiceImpl service_;
+  std::unique_ptr<grpc::Server> server_;
+  std::unique_ptr<EchoTestService::Stub> stub_;
+};
+
 TEST_F(OTelPluginEnd2EndTest, ClientAttemptStarted) {
-  Init({grpc::internal::OTelClientAttemptStartedInstrumentName()});
+  SetUp({grpc::internal::OTelClientAttemptStartedInstrumentName()});
   SendRPC();
   const char* kMetricName = "grpc.client.attempt.started";
   auto data = ReadCurrentMetricsData(
@@ -74,7 +185,7 @@ TEST_F(OTelPluginEnd2EndTest, ClientAttemptStarted) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ClientAttemptDuration) {
-  Init({grpc::internal::OTelClientAttemptDurationInstrumentName()});
+  SetUp({grpc::internal::OTelClientAttemptDurationInstrumentName()});
   SendRPC();
   const char* kMetricName = "grpc.client.attempt.duration";
   auto data = ReadCurrentMetricsData(
@@ -105,8 +216,8 @@ TEST_F(OTelPluginEnd2EndTest, ClientAttemptDuration) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ClientAttemptSentTotalCompressedMessageSize) {
-  Init({grpc::internal::
-            OTelClientAttemptSentTotalCompressedMessageSizeInstrumentName()});
+  SetUp({grpc::internal::
+             OTelClientAttemptSentTotalCompressedMessageSizeInstrumentName()});
   SendRPC();
   const char* kMetricName =
       "grpc.client.attempt.sent_total_compressed_message_size";
@@ -138,8 +249,8 @@ TEST_F(OTelPluginEnd2EndTest, ClientAttemptSentTotalCompressedMessageSize) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ClientAttemptRcvdTotalCompressedMessageSize) {
-  Init({grpc::internal::
-            OTelClientAttemptRcvdTotalCompressedMessageSizeInstrumentName()});
+  SetUp({grpc::internal::
+             OTelClientAttemptRcvdTotalCompressedMessageSizeInstrumentName()});
   SendRPC();
   const char* kMetricName =
       "grpc.client.attempt.rcvd_total_compressed_message_size";
@@ -171,7 +282,7 @@ TEST_F(OTelPluginEnd2EndTest, ClientAttemptRcvdTotalCompressedMessageSize) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ServerCallStarted) {
-  Init({grpc::internal::OTelServerCallStartedInstrumentName()});
+  SetUp({grpc::internal::OTelServerCallStartedInstrumentName()});
   SendRPC();
   const char* kMetricName = "grpc.server.call.started";
   auto data = ReadCurrentMetricsData(
@@ -195,7 +306,7 @@ TEST_F(OTelPluginEnd2EndTest, ServerCallStarted) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ServerCallDuration) {
-  Init({grpc::internal::OTelServerCallDurationInstrumentName()});
+  SetUp({grpc::internal::OTelServerCallDurationInstrumentName()});
   SendRPC();
   const char* kMetricName = "grpc.server.call.duration";
   auto data = ReadCurrentMetricsData(
@@ -222,8 +333,8 @@ TEST_F(OTelPluginEnd2EndTest, ServerCallDuration) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ServerCallSentTotalCompressedMessageSize) {
-  Init({grpc::internal::
-            OTelServerCallSentTotalCompressedMessageSizeInstrumentName()});
+  SetUp({grpc::internal::
+             OTelServerCallSentTotalCompressedMessageSizeInstrumentName()});
   SendRPC();
   const char* kMetricName =
       "grpc.server.call.sent_total_compressed_message_size";
@@ -251,8 +362,8 @@ TEST_F(OTelPluginEnd2EndTest, ServerCallSentTotalCompressedMessageSize) {
 }
 
 TEST_F(OTelPluginEnd2EndTest, ServerCallRcvdTotalCompressedMessageSize) {
-  Init({grpc::internal::
-            OTelServerCallRcvdTotalCompressedMessageSizeInstrumentName()});
+  SetUp({grpc::internal::
+             OTelServerCallRcvdTotalCompressedMessageSizeInstrumentName()});
   SendRPC();
   const char* kMetricName =
       "grpc.server.call.rcvd_total_compressed_message_size";
@@ -281,10 +392,8 @@ TEST_F(OTelPluginEnd2EndTest, ServerCallRcvdTotalCompressedMessageSize) {
 
 // Make sure that no meter provider results in normal operations.
 TEST_F(OTelPluginEnd2EndTest, NoMeterProviderRegistered) {
-  Init({grpc::internal::OTelClientAttemptStartedInstrumentName()},
-       /*resource=*/opentelemetry::sdk::resource::Resource::Create({}),
-       /*labels_injector=*/nullptr,
-       /*test_no_meter_provider=*/true);
+  SetUp({grpc::internal::OTelClientAttemptStartedInstrumentName()},
+        /*test_no_meter_provider=*/true);
   SendRPC();
 }
