+++ tools/distrib/fix_build_deps.py
@@ -94,10 +94,6 @@ EXTERNAL_DEPS = {
     "absl/utility/utility.h": "absl/utility",
     "address_sorting/address_sorting.h": "address_sorting",
     "google/cloud/opentelemetry/resource_detector.h": "google_cloud_cpp:experimental-opentelemetry",
-    "opentelemetry/common/attribute_value.h": "otel/api",
-    "opentelemetry/common/key_value_iterable.h": "otel/api",
-    "opentelemetry/nostd/function_ref.h": "otel/api",
-    "opentelemetry/nostd/string_view.h": "otel/api",
     "opentelemetry/context/context.h": "otel/api",
     "opentelemetry/metrics/meter.h": "otel/api",
     "opentelemetry/metrics/meter_provider.h": "otel/api",
@@ -106,8 +102,6 @@ EXTERNAL_DEPS = {
     "opentelemetry/nostd/shared_ptr.h": "otel/api",
     "opentelemetry/nostd/unique_ptr.h": "otel/api",
     "opentelemetry/sdk/metrics/meter_provider.h": "otel/sdk/src/metrics",
-    "opentelemetry/sdk/common/attribute_utils.h": "otel/sdk:headers",
-    "opentelemetry/sdk/resource/semantic_conventions.h": "otel/sdk:headers",
     "ares.h": "cares",
     "fuzztest/fuzztest.h": ["fuzztest", "fuzztest_main"],
     "google/api/monitored_resource.pb.h": (
